+++ CHANGELOG.md
@@ -21,6 +21,10 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Added `HashMap.forEach()`
+
 ### Dependencies
 
 * Updated dev dependencies to resolve low-severity `npm audit` warnings.